requirements.txt
streamlit
autogen-core
autogen-ext[openai]
plotly
pandas
pyyaml